Rename debug_tethys to debug_tethys_core
