+++ CMakeLists.txt
@@ -126,6 +126,6 @@ add_sanitizers(sophos_debug)
 add_sanitizers(diana_debug)
 add_sanitizers(janus_debug)
 add_sanitizers(oceanus_debug)
-add_sanitizers(tethys_debug)
+add_sanitizers(tethys_core_debug)
 
 coverage_evaluate()
